+++ database/migrations/0001_01_01_000001_create_cache_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (! app()->environment(['local', 'testing'])) {
+            return;
+        }
+
         Schema::create('cache', function (Blueprint $table) {
             $table->string('key')->primary();
             $table->mediumText('value');
@@ -29,6 +33,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (! app()->environment(['local', 'testing'])) {
+            return;
+        }
+
         Schema::dropIfExists('cache');
         Schema::dropIfExists('cache_locks');
     }
